Random try to fix strange error
Backend and Neo4j base images are built fine but Github fails to build
`humanconnection/webapp` docker image. Is a request made to dockerhub to
see if the repository exists? Even if the resulting image is not being
pushed to dockerhub?

What really bugs me is that `humanconnection/backend` does not exist
anymore and yet the Github actions runs fine building that image.