release: also update helmchart versions
To have congruence a release also updates the helmchart version. This
utilizes sed to replace the coresponding string. The `sed` command is
not present on windows and could prevent a windows user to use the
release functionality out of the box.

The release functionality is also externalized into its own script
`scripts/release.sh`.