refactor: display server errors on the logs
So, this might clutter our CI but I found it so helpful to turn this on.
Our seeds are riddled with runtime errors that we never see. It's also
easier to debug the tests if you can see the actual errors thrown by the
backend.

Let's see how CI log output looks like after this.