refactor(backend): test image and attachment urls
This is a leftover of #8558. I wanted to test-drive a previous solution where I would replace the domain of the returned S3 `Location`. I found a cleaner solution with a local proxy, so the test is not required anymore.

However I thought it might have some value. At least it's testing the file attribute of attachments and we don't have a test for that yet.